Added example of Vega-Lite visualisations
Vega-Lite is an higher-level syntax to define visualisations.
In this example the Vega-Lite spec is converted into a Vega spec which
is then rendered.

NOTE: In this example the Vega-Lite spec is hardcoded and the data is
also embedded into the specification (e.g. they're not read from S3).
The point of this example is to show how to convert/render a Vega-Lite
visualisation, but of course this and its data can come from S3 as in
the 2nd example.

I've also added some documentation and types.