Add helper Docker images … which can be built by any CI deployment pipeline as they are intended to be standalone and non-configurable at build time.
These helper images do not form part of the application, but rather provide support functions:

- `default-backend`: This image serves simple, static and fully self-contained error and maintenance pages.
Its primary use is as the default backend for the application’s kubernetes ingress to respond in case particular
HTTP error codes are produced by the application/ingress. For example, if the connecting client is forbidden due
to IP address restrictions, an informative page can be shown in place of the generic "403 Forbidden" message
served by the default nginx ingress controller.
- `port-forward`: This small image can be used to forward TCP services only accessible from within the cluster.
For example, it can be used to connect to Elasticache Redis via a local port.