ESLint: prettier/prettier's problems as warnings
If something is syntactically valid but subjectively not pretty it should
be highlighted in a different/less alarming way (e.g. yellow squiggly line
instead of red).

Before this it was very distracting/annoying to see red squiggly lines
in the editor and then discover (by hovering on the "error") that
it was just a style problem (that can possibly wait).