Removed dependency used to generate random data CSVs
The objects-to-csv package didn't have any recent releases and convert
an array of obejcts to CSV is relatively straightforward.

By default this command generates CSVs with ~80,000 items, and this was
not a problem before. The data is all kept in memory as before but after
this change it's possible to generate much bigger CSV files if necessary.